LayoutTests/webgpu/command-buffers-expected.txt

PASS [object WebGPU] is defined.
PASS Successfully created WebGPUCommandBuffer.
All tests complete.
PASS successfullyParsed is true

TEST COMPLETE
